requirements.txt
certifi==2022.12.7
cffi==1.15.1
circuitbreaker==1.4.0
cryptography==39.0.2
oci==2.99.1
pycparser==2.21
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
